requirements.txt
django==1.9.6
djangorestframework==3.6.2